+++ R/forest.rma.r
@@ -1302,12 +1302,12 @@ lty, fonts, cex, cex.lab, cex.axis, ...) {
             sel <- xs <= beta.pi.lb
             xs.sel <- xs[sel]
             ys.sel <- ys[sel]
-            lpolygon(c(xs.sel,rev(xs.sel)), c(ys.sel,rep(drow,length(ys.sel))), col=col[2], border=border[2], ...)
+            lpolygon(c(xs.sel,rev(xs.sel)), c(ys.sel,rep(drow,length(ys.sel))), col=col[2], border=ifelse(is.na(col[2]), NA, border[2]), ...)
 
             sel <- xs >= beta.pi.ub
             xs.sel <- xs[sel]
             ys.sel <- ys[sel]
-            lpolygon(c(xs.sel,rev(xs.sel)), c(ys.sel,rep(drow,length(ys.sel))), col=col[2], border=border[2], ...)
+            lpolygon(c(xs.sel,rev(xs.sel)), c(ys.sel,rep(drow,length(ys.sel))), col=col[2], border=ifelse(is.na(col[2]), NA, border[2]), ...)
 
             ### add horizontal and distribution lines